requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.7
dlib==19.24.2
filelock==3.12.4
fsspec==2023.9.2
idna==3.4
image==1.5.33
imutils==0.5.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2
numpy==1.26.1
opencv-python==4.8.1.78
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
torch==2.1.0
torchvision==0.16.0
typing_extensions==4.8.0
urllib3==2.0.7
